# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SaleSync API",
        "description": "API documentation for SaleSync",
        "version": "1.0.0"
    },
    "paths": {
        "api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "8be708a7cdb6f251b6fb8fbbd41b3b92",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User e-mail",
                                        "type": "string",
                                        "example": "example@example"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "operationId": "503df0f0f91e269dbfb90be06b4fd408",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "User name",
                                        "type": "string",
                                        "example": "Test User"
                                    },
                                    "email": {
                                        "description": "User e-mail",
                                        "type": "string",
                                        "example": "example@example"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "User password confirmation",
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get all products",
                "description": "Get all products",
                "operationId": "5afdc398c410dcbc922743e99c89272b",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create a new product",
                "description": "Create a new product",
                "operationId": "8a4f16a85b1b86d40798a70e6667b405",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "category_id",
                                    "sku",
                                    "price",
                                    "stock"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Product name",
                                        "type": "string",
                                        "example": "Product Test"
                                    },
                                    "description": {
                                        "description": "Product description",
                                        "type": "string",
                                        "example": "Product description"
                                    },
                                    "category_id": {
                                        "description": "Product category id",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "sku": {
                                        "description": "Product sku",
                                        "type": "string",
                                        "example": "product-test"
                                    },
                                    "price": {
                                        "description": "Product price",
                                        "type": "number",
                                        "example": 10
                                    },
                                    "stock": {
                                        "description": "Product stock",
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product-detail": {
            "get": {
                "tags": [
                    "Product"
                ],
                "description": "Returns the details of a product",
                "operationId": "5a3b16604c5bf2aedb50ece076ae7aee",
                "parameters": [
                    {
                        "parameter": "sku",
                        "name": "sku",
                        "in": "query",
                        "description": "Product sku",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}